+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `ApiControllerConcern#render_json_error` improvements
 - Add Agilibox::Service object
 - Fix SMS AmazonSNS strategy bug
+- Fix time period filter
 
 ## 1.0.14